refactor: rename emscripten wasm file
