+++ .prettierignore
@@ -2,4 +2,4 @@ package-lock.json
 modules/
 dist
 coverage
-emscripten-wasm.js
+EmscriptenWasm.js
