+++ src/mpg123-decoder/src/MPEGDecoderWebWorker.js
@@ -1,6 +1,6 @@
 import Worker from "web-worker";
 
-import EmscriptenWASM from "./emscripten-wasm.js";
+import EmscriptenWASM from "./EmscriptenWasm.js";
 import MPEGDecodedAudio from "./MPEGDecodedAudio.js";
 import MPEGDecoder from "./MPEGDecoder.js";
