+++ src/ogg-opus-decoder/src/EmscriptenWasm.js
