+++ src/opus-decoder/src/EmscriptenWasm.js
