+++ src/opus-decoder/src/OpusDecoderWebWorker.js
@@ -1,5 +1,5 @@
 import Worker from "web-worker";
-import EmscriptenWASM from "./emscripten-wasm.js";
+import EmscriptenWASM from "./EmscriptenWasm.js";
 import OpusDecodedAudio from "./OpusDecodedAudio.js";
 import OpusDecoder from "./OpusDecoder.js";
